UCT/IB: Retry ODP registration unlimitedly

+++ src/uct/ib/base/ib_md.h
@@ -94,7 +94,7 @@ typedef struct uct_ib_md_ext_config {
                                                        invalidated memory keys
                                                        that are kept idle before
                                                        reuse*/
-    unsigned                 reg_retry_cnt; /**< Memory registration retry count */
+    unsigned long            reg_retry_cnt; /**< Memory registration retry count */
     unsigned                 smkey_block_size; /**< Mkey indexes in a symmetric block */
 } uct_ib_md_ext_config_t;
